backend/requirements.txt
Django==5.1.2
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
scipy==1.13.1
scikit-learn==1.5.0
spotipy==2.24.0
umap-learn==0.5.7
plotly==5.24.0
whitenoise==6.8.2
fontawesomefree==6.6.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
torch==2.5.1
transformers==4.46.3
deezer-python==7.1.0
librosa==0.10.2
soundfile==0.12.1